Improve Google OAuth login/register flow
