+++ src/containers/account/login.tsx
@@ -1,19 +1,5 @@
-import { useRouter } from 'next/router';
-import React, { useEffect } from 'react';
-
 import withAuthGuard from '@components/HOCs/withAuthGuard';
-import { makeDiscoverRoute } from '@lib/routes';
-import useLanguageRoute from '@lib/useLanguageRoute';
 
-function Login(): JSX.Element {
-	const router = useRouter();
-	const route = useLanguageRoute();
+import LoginRedirect from './loginRedirect';
 
-	useEffect(() => {
-		router.push(makeDiscoverRoute(route));
-	}, [router, route]);
-
-	return <>redirecting...</>;
-}
-
-export default withAuthGuard(Login);
+export default withAuthGuard(LoginRedirect);
